+++ test/app/GET/ui/config/admin-ui/imprint.de.html
@@ -1,4 +1,4 @@
-<h1>Impressum</h1>
+<h1>Muster Impressum</h1>
 
 <p>
 <b>Test Universität</b><br>
