+++ test/app/GET/ui/config/admin-ui/imprint.en-GB.html
@@ -1,4 +1,4 @@
-<h1>Imprint</h1>
+<h1>Sample Imprint</h1>
 
 <p>
 <b>Test University</b><br>
