+++ test/app/GET/ui/config/admin-ui/privacy.de.html
@@ -1,11 +1,1 @@
-<h1>Datenschutzerklärung</h1>
-
-<p>
-<ul>
-    <li>Name</li>
-    <li>Email</li>
-    <ul>
-        <li>test</li>
-    </ul>
-</ul>
-</p>
+<h1>Muster Datenschutzerklärung</h1>
