+++ test/app/GET/ui/config/admin-ui/privacy.en-GB.html
@@ -1,8 +1,1 @@
-<h1>Privacy statement</h1>
-
-<p>
-<ul>
-    <li>Name</li>
-    <li>Email</li>
-</ul>
-</p>
+<h1>Sample Privacy statement</h1>
